powersj scan results of telegraf image
This was produces as part of the final project for the Chainguard
Vulnerability Management-Certification. I used the Telegraf Official
Docker image with a custom Telegraf config and scanned the image. That
image is based off a Debian image.

Then I switched to using the Chainguard Telegraf image. While I was
hoping to produce an image with smartmontools installed, it seems wolfi
does not have that package yet, so I could not multi-stage build. But
this clearly shows the removal of the OS CVEs and only the potential Go
binary CVEs remain.